+++ client/src/scripts/others/config.ts
@@ -27,7 +27,11 @@ export const Debug={
 
 export const ConfigCasters=Object.freeze({
     sv_loadout_name:Casters.toString,
-    sv_loadout_skin:Casters.toString,
+    sv_loadout_female:Casters.toBoolean,
+    sv_loadout_body_tint:Casters.toString,
+    sv_loadout_hair:Casters.toString,
+    sv_loadout_hair_tint:Casters.toString,
+    sv_loadout_shirt:Casters.toString,
 
     sv_graphics_resolution:Casters.generateUnionCaster(["low","medium"]),
     sv_graphics_renderer:Casters.generateUnionCaster(["webgl1","webgl2"]),
@@ -57,16 +61,14 @@ export const ConfigCasters=Object.freeze({
     sv_ui_special_color:Casters.toString,
     sv_ui_translation:Casters.toString,
     sv_ui_interactive:Casters.toBoolean,
-
-    sv_loadout_female:Casters.toBoolean,
-    sv_loadout_body_tint:Casters.toString,
-    sv_loadout_hair:Casters.toString,
-    sv_loadout_hair_tint:Casters.toString,
-    sv_loadout_shirt:Casters.toString,
 })
 export const ConfigDefaultValues={
-    sv_loadout_skin:"default_skin",
     sv_loadout_name:"",
+    sv_loadout_female:false,
+    sv_loadout_body_tint:"#f0a93f",
+    sv_loadout_hair:"hair_1",
+    sv_loadout_hair_tint:"#222222",
+    sv_loadout_shirt:"blue_shirt",
 
     sv_graphics_renderer:"webgl2",
     sv_graphics_resolution:(Debug.force_mobile||isMobile)?"low":"medium",
@@ -97,12 +99,6 @@ export const ConfigDefaultValues={
     sv_ui_special_color:"#fffb00",
     sv_ui_translation:"en",
     sv_ui_interactive:true,
-
-    sv_loadout_female:false,
-    sv_loadout_body_tint:"#f0a93f",
-    sv_loadout_hair:"hair_1",
-    sv_loadout_hair_tint:"#222222",
-    sv_loadout_shirt:"blue_shirt",
 }
 export const ConfigDefaultActions={
     "move_up":{