+++ common/engine/client/misc/keys.ts
@@ -185,10 +185,12 @@ export enum InputEventType {
 export type InputKeyEvent = {
     type: InputEventType.KeyDown | InputEventType.KeyUp
     key: number
+    element?:HTMLElement
 }
 export type InputActionEvent = {
     type: InputEventType.ActionDown | InputEventType.ActionUp
     action: string
+    element?:HTMLElement
 }
 export type InputAxisEvent = {
     type: InputEventType.Axis
@@ -211,6 +213,7 @@ export class InputManager {
     pressed = new Set<number>()
     down = new Set<number>()
     up = new Set<number>()
+    key_elements:Record<number,HTMLElement>={}
 
     gamepad_pressed = new Set<number>()
     private wheel_pressed = new Set<number>()
@@ -239,69 +242,76 @@ export class InputManager {
     }
     bind(canvas: HTMLCanvasElement,elem: HTMLElement = document.body) {
         elem.tabIndex = 1
-        elem.addEventListener("keydown",this.on_key_down)
-        elem.addEventListener("keyup",this.on_key_up)
-        canvas.addEventListener("mousedown",this.on_mouse_down)
-        elem.addEventListener("mouseup",this.on_mouse_up)
-        elem.addEventListener("wheel",this.on_wheel,{passive: false})
+        elem.addEventListener("keydown",this.on_key_down())
+        elem.addEventListener("keyup",this.on_key_up())
+        elem.addEventListener("mousedown",this.on_mouse_down())
+        elem.addEventListener("mouseup",this.on_mouse_up())
+        elem.addEventListener("wheel",this.on_wheel(),{passive: false})
         elem.addEventListener("pointermove",e => this.on_pointer_move(e, canvas))
     }
     private emit(event: InputEvent) {
         this.listener.emit(event.type, event)
     }
-    private on_key_down = (e: KeyboardEvent) => {
-        if (!this.focus) return
-        const key = KeyNames[e.keyCode]
-        if (key === undefined) return
-        if (!this.pressed.has(key)) {
-            this.down.add(key)
-            this.emit({type: InputEventType.KeyDown,key})
+    private on_key_down(){
+        return (e: KeyboardEvent) => {
+            if (!this.focus) return
+            const key = KeyNames[e.keyCode]
+            if (key === undefined) return
+            if (!this.pressed.has(key)) {
+                this.down.add(key)
+                this.emit({type: InputEventType.KeyDown,key,element:e.target as HTMLElement})
+            }
+            this.pressed.add(key)
         }
-        this.pressed.add(key)
     }
-    private on_key_up = (e: KeyboardEvent) => {
-        const key = KeyNames[e.keyCode]
-        if (key === undefined) return
-        this.pressed.delete(key)
-        this.up.add(key)
-        this.emit({type: InputEventType.KeyUp,key})
-    }
-    private on_mouse_down = (e: MouseEvent) => {
-        if (!this.focus) return
-        const key = KeyNames[e.button + 1000]
-        if (key === undefined) return
-        if (!this.pressed.has(key)) {
-            this.down.add(key)
-            this.emit({type: InputEventType.KeyDown,key})
+    private on_key_up(){
+        return (e: KeyboardEvent) => {
+            const key = KeyNames[e.keyCode]
+            if (key === undefined) return
+            this.pressed.delete(key)
+            this.up.add(key)
+            this.emit({type: InputEventType.KeyUp,key,element:e.target as HTMLElement})
         }
-        this.pressed.add(key)
     }
-    private on_mouse_up = (e: MouseEvent) => {
-        const key = KeyNames[e.button + 1000]
-        if (key === undefined) return
-        this.pressed.delete(key)
-        this.up.add(key)
-        this.emit({type: InputEventType.KeyUp,key})
-    }
-    private on_wheel = (e: WheelEvent) => {
-        if (!this.focus) {
-            return
+    private on_mouse_down(){
+        return (e: MouseEvent) => {
+            if (!this.focus) return
+            const key = KeyNames[e.button + 1000]
+            if (key === undefined) return
+            if (!this.pressed.has(key)) {
+                this.down.add(key)
+                this.emit({type: InputEventType.KeyDown,key,element:e.target as HTMLElement})
+                this.key_elements[key]=e.target as HTMLElement
+            }
+            this.pressed.add(key)
         }
-        const key=e.deltaY<0?Key.Mouse_Wheel_Up:Key.Mouse_Wheel_Down
-        this.wheel_pressed.add(key)
+    }
+    private on_mouse_up(){
+        return (e: MouseEvent) => {
+            const key = KeyNames[e.button + 1000]
+            if (key === undefined) return
+            this.pressed.delete(key)
+            this.up.add(key)
+            this.emit({type: InputEventType.KeyUp,key,element:e.target as HTMLElement})
+            delete this.key_elements[key]
+        }
+    }
+    private on_wheel(){
+        return (e: WheelEvent) => {
+            if (!this.focus) {
+                return
+            }
+            const key=e.deltaY<0?Key.Mouse_Wheel_Up:Key.Mouse_Wheel_Down
+            this.wheel_pressed.add(key)
 
-        this.pressed.add(key)
-        this.down.add(key)
-        this.up.add(key)
+            this.pressed.add(key)
+            this.down.add(key)
+            this.up.add(key)
 
-        this.emit({
-            type: InputEventType.KeyDown,
-            key
-        })
-        this.emit({
-            type: InputEventType.KeyUp,
-            key
-        })
+            this.emit({type: InputEventType.KeyDown,key,element:e.target as HTMLElement})
+            this.emit({type: InputEventType.KeyUp,key,element:e.target as HTMLElement})
+            this.key_elements[key]=e.target as HTMLElement
+        }
     }
     private on_pointer_move(e: PointerEvent,canvas: HTMLCanvasElement) {
         if (!this.focus) return
@@ -403,23 +413,23 @@ export class InputManager {
     keyUp(key: number): boolean {
         return this.up.has(key)
     }
-    action_pressed(action: InputAction): boolean {
+    action_pressed(action: InputAction): {key:number,button:number}|undefined {
         for (const k of action.keys) {
             if(this.pressed.has(k)) {
-                return true
+                return {key:k,button:-1}
             }
         }
         for (const b of action.buttons) {
             if(this.gamepad_pressed.has(b)) {
-                return true
+                return {button:b,key:-1}
             }
         }
-        return false
+        return
     }
     action_id_pressed(id: string): boolean {
         const action = this.actions[id]
         if (!action) return false
-        return this.action_pressed(action)
+        return this.action_pressed(action)!==undefined
     }
     wait_for_action(action: string): Promise<void> {
         return new Promise((resolve) => {
@@ -499,13 +509,14 @@ export class InputManager {
                 this.active_actions.add(action)
                 this.emit({
                     type: InputEventType.ActionDown,
-                    action
+                    action,
+                    element:this.key_elements[pressed.key]
                 })
             }else if(!pressed&&active){
                 this.active_actions.delete(action)
                 this.emit({
                     type: InputEventType.ActionUp,
-                    action
+                    action,
                 })
             }
         }